+++ src/screens/WalletConnect/Home/WalletConnectHome.js
@@ -51,7 +51,7 @@ import { firebaseRemoteConfig } from 'services/firebase';
 import { mapNotNil } from 'utils/array';
 import { appFont, fontStyles, spacing } from 'utils/variables';
 import { useChainsConfig } from 'utils/uiConfig';
-import { openUrl } from 'utils/inAppBrowser';
+import { openUrl, showServiceLaunchError } from 'utils/inAppBrowser';
 
 // Types
 import type { SectionBase } from 'utils/types/react-native';
@@ -106,6 +106,8 @@ function WalletConnectHome() {
       } else {
         Linking.openURL(url);
       }
+    } else {
+      showServiceLaunchError();
     }
   };
